+++ src/Responses/LaravelResponseFactory.php
@@ -4,6 +4,4 @@ namespace Axn\LaravelGlide\Responses;
 
 use League\Glide\Responses\SymfonyResponseFactory;
 
-class LaravelResponseFactory extends SymfonyResponseFactory
-{
-}
+class LaravelResponseFactory extends SymfonyResponseFactory {}